requirements.txt
phidata 
python-dotenv
duckduckgo-search
streamlit
google-generativeai
PyPDF2